requirements.txt
playwright
pyyaml
textblob
pandas
